Reorganize things so tests can actually be run
