+++ otherdave/commands/__init__.py
@@ -1,5 +1,4 @@
 # otherdave/commands/__init__.py
 
-from .dlog import *
 from .haiku import *
 from .update import *
