+++ otherdave/util/__init__.py
@@ -1,0 +1,3 @@
+# /otherdave/otherdave/util/__init__.py
+
+from .dlog import *
