+++ test/commands/__init__.py
@@ -1,3 +1,1 @@
 # otherdave/test/commands/__init__.py
-
-from .test_haiku import test_haiku
